Migrate portfolio helper to TS
